Provide non-implementation specific CopyOptions
According to the java8 docs:
https://docs.oracle.com/javase/8/docs/api/java/nio/file/CopyOption.html
`CopyOption` is an interface that both `LinkOption` and
`StandardCopyOption` implement. As one example, this behavior is used in
`Files.copy(Path,Path,CopyOption...)`.

In the better-files code, this is fixed to two separate types:

```scala
type CopyOptions = Seq[StandardCopyOptions]
type LinkOptions = Seq[LinkOption]
```

This is troublesome, since, let's say we want to change the
implementation of `copyTo` to include `LinkOption.NOFOLLOW_LINKS` via
`implicit` parameters it wouldn't work.

This fixes the problem by making the current `CopyOptions` be
`StandardCopyOptions` type and a new `CopyOptions` type that will
include both `StandardCopyOption` and `LinkOption`

Ref: #93