glob: add support for automatic path prefix
Support automatic path prefix with glob (`pathGlob` and `pathRegex`).
In this mode, pattern is prefixed with `File`-instance's path.

Standard java PatternMatch names `regex` and `glob` behaves
as they do in java world.

Default is `pathGlob`, that default behaviour is similar
with  Unix `ls` or with
[Python 3.5+ glob](https://docs.python.org/3/library/glob.html)

Please see #114 for background information.