Implement File.resource and resourceAsStream with macros, to avoid failure in some situations.
ClassLoader#getResource[AsStream] is caller-sensitive: access to the resource is determined in part by where the methods are called from. This is especially important in the presence of Java ≥ 9 modules, because access is denied by default to resources (except class files and META-INF) in other modules. Therefore, calling these methods through a library like better-files from inside a Java 9+ module will usually fail.

Implementing these methods as macros should solve the problem, by inlining the getResource[AsStream] calls in the modules they're really supposed to be called from.

Note: This commit deprecates File.copyResource, and leaves its behavior unchanged. That method is impossible to implement as a macro, because of the default parameter value in its second parameter list. (Macros are not allowed to have default parameter values.) The deprecation message explains what to do instead.

Note: Dotty's new inline behavior would make this change unnecessary. Unfortunately, while Scala 2 has an @inline annotation, the compiler is not required to honor it, and doesn't by default. The only way to force inlining in Scala 2 is with a macro.