Teach Resource how to look up resources from other places: different class loaders, specific classes, or the calling class. Move tests into new class ResourceSpec.
Specifically, add four new methods to Resource:
· from — look up resources using a different ClassLoader
· at — look up resources relative to a specified class (two variants)
· my — look up resources relative to the caller class
Tests for all of these (along with the original Resource tests) are in the new class ResourceSpec.

This commit uses a bit of a hack to make such “modifier” methods work with macros, but it works, and the resulting API is easy to use. See Resource.scala line 33 for an explanation.

This commit adds an intransitive dependency on ghik/silencer, a compiler plugin+annotation to selectively suppress warnings (like Java @SuppressWarnings). This is needed because:
· scalac generates bogus warnings of unused pattern variables, when the pattern contains a quasiquote.
· Resource.{from,at,my} are stubs; their parameters are used externally, but not by the method itself.
This dependency is in the “provided,optional” configuration. It appears in the generated POM, but other projects depending on better.files won't inherit that dependency. (It would be more correct to use the “compile-internal,test-internal” configuration, but IntelliJ doesn't understand such dependencies. This is the next best.)

The existing dependency on scala.reflect now also has “provided,optional” configuration.